requirements.txt
pinecone-client
streamlit
sentence-transformers
openai
